securities_lending_optimizer/requirements.txt
pandas>=2.0
numpy>=1.24
lightgbm>=4.0
xgboost>=2.0
scikit-learn>=1.3
cvxpy>=1.4
ecos>=2.0
stable-baselines3>=2.2
gymnasium>=0.29
pymoo>=0.6
torch>=2.0
optuna>=3.4
shap>=0.44
requests>=2.31
pyyaml>=6.0
python-dotenv>=1.0
joblib>=1.3
